feat: array of games per block in prover
