+++ contracts/script/deploy/config/devnet/leaderboard.json
@@ -7,7 +7,7 @@
 	"levels": [
 		{
 			"endsAtBlock": "0x000000000000000000000000000000000000000000000000000000174876e800",
-			"gameConfig": "0xfa0a0300fa0c0500c8140c00b41e0a01aa1e0a03a01e0a07a0320501a0640c08",
+			"gameConfig": "0xff0a0300fa0c0500c8140c00b41e0a01aa1e0a03a01e0a07a0320501a0640c08",
 			"startsAtBlock": "0x0000000000000000000000000000000000000000000000000000000000000000"
 		}
 	]
