+++ games/beast/beast1984/src/game.rs
@@ -1,14 +1,19 @@
 //! this module contains the main struct that orchestrates the game
 
-use crate::{ethereum, help::Help, prover::{prove, save_proof}, stty::{RawMode, install_raw_mode_signal_handler}};
+use crate::{
+    ethereum,
+    help::Help,
+    prover::{prove, save_proof},
+    stty::{install_raw_mode_signal_handler, RawMode},
+};
 use game_logic::{
-    ANSI_BOLD, ANSI_LEFT_BORDER, ANSI_RESET, ANSI_RESET_BG, ANSI_RESET_FONT, ANSI_RIGHT_BORDER,
-    BOARD_HEIGHT, BOARD_WIDTH, Dir, LOGO, Tile,
     beasts::{Beast, BeastAction, CommonBeast, Egg, HatchedBeast, HatchingState, SuperBeast},
     board::Board,
-    common::levels::Level,
+    common::{game::GameMatch, levels::Level},
     player::{Player, PlayerAction},
     proving::{GameLogEntry, LevelLog},
+    Dir, Tile, ANSI_BOLD, ANSI_LEFT_BORDER, ANSI_RESET, ANSI_RESET_BG, ANSI_RESET_FONT,
+    ANSI_RIGHT_BORDER, BOARD_HEIGHT, BOARD_WIDTH, LOGO,
 };
 use std::{
     io::{self, Read},
@@ -78,10 +83,12 @@ pub enum GameState {
 
 /// this is our main game struct that orchestrates the game and its bits
 pub struct Game {
+    pub block_number: u64,
     /// our board
     pub board: Board,
     /// the current level we're in
     pub level: Level,
+    pub game_match: GameMatch,
     /// when we started the level
     pub level_start: Instant,
     /// all of our common beasts instances
@@ -108,10 +115,11 @@ pub struct Game {
 impl Game {
     /// create a new instance of the beast game
     pub fn new() -> Self {
-
         let address = ethereum::read_address();
-
-        let board_terrain_info = Board::generate_terrain(Level::One);
+        // TODO: get block number from eth rpc
+        let block_number = 0;
+        let game_match = GameMatch::new(block_number);
+        let board_terrain_info = Board::generate_terrain(game_match.get_config(Level::One));
 
         install_raw_mode_signal_handler();
         let _raw_mode = RawMode::enter().unwrap_or_else(|error| {
@@ -142,8 +150,10 @@ impl Game {
 
         Self {
             levels_completion_log: vec![fist_level_log],
+            block_number: 0,
             board: Board::new(board_terrain_info.buffer),
             level: Level::One,
+            game_match,
             level_start: Instant::now(),
             common_beasts: board_terrain_info.common_beasts,
             super_beasts: board_terrain_info.super_beasts,
@@ -160,7 +170,7 @@ impl Game {
     }
 
     pub fn start_new_game(&mut self) {
-        let board_terrain_info = Board::generate_terrain(Level::One);
+        let board_terrain_info = Board::generate_terrain(self.game_match.get_config(self.level));
         let board = Board::new(board_terrain_info.buffer);
         let fist_level_log = LevelLog {
             level: Level::One,
@@ -360,8 +370,8 @@ impl Game {
             }
 
             // eggs hatching
-            self.eggs
-                .retain_mut(|egg| match egg.hatch(self.level.get_config()) {
+            self.eggs.retain_mut(
+                |egg| match egg.hatch(self.game_match.get_config(self.level)) {
                     HatchingState::Incubating => true,
                     HatchingState::Hatching(position, _instant) => {
                         self.board[&position] = Tile::EggHatching;
@@ -372,7 +382,8 @@ impl Game {
                         self.board[&position] = Tile::HatchedBeast;
                         false
                     }
-                });
+                },
+            );
 
             // end game through no more beasts
             if self.common_beasts.len()
@@ -515,7 +526,7 @@ impl Game {
         let _ = handle.join();
 
         if let Some(level) = self.level.next() {
-            let board_terrain_info = Board::generate_terrain(level);
+            let board_terrain_info = Board::generate_terrain(self.game_match.get_config(level));
             let board = Board::new(board_terrain_info.buffer);
 
             let level_log = LevelLog {
@@ -532,7 +543,7 @@ impl Game {
             self.eggs = board_terrain_info.eggs;
             self.hatched_beasts = board_terrain_info.hatched_beasts;
             self.player.position = board_terrain_info.player.position;
-            self.player.score += self.level.get_config().completion_score;
+            self.player.score += self.game_match.get_config(self.level).completion_score;
             self.state = GameState::Playing;
         } else {
             self.has_won = true;
@@ -603,8 +614,9 @@ impl Game {
         };
 
         let address = self.address.clone();
+        let block_number = self.block_number.clone();
         let handle = std::thread::spawn(move || {
-            let res = prove(levels_completion_log, address);
+            let res = prove(levels_completion_log, block_number, address);
             if let Ok(receipt) = res {
                 save_proof(receipt).expect("To be able to write proof");
             } else {
@@ -633,7 +645,7 @@ impl Game {
 
     fn get_secs_remaining(&self) -> u64 {
         let elapsed = Instant::now().duration_since(self.level_start);
-        let total_time = self.level.get_config().time;
+        let total_time = self.game_match.get_config(self.level).time;
         if total_time > elapsed {
             total_time - elapsed
         } else {
