+++ games/beast/game_logic/src/common/mod.rs
@@ -1,1 +1,2 @@
+pub mod game;
 pub mod levels;
