+++ games/beast/game_logic/src/proving.rs
@@ -22,5 +22,6 @@ pub struct LevelLog {
 #[derive(Serialize, Deserialize)]
 pub struct ProgramInput {
     pub levels_log: Vec<LevelLog>,
+    pub block_number: u64,
     pub address: Vec<u8>,
 }
